% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAvailable.R
\name{listExistingObjects}
\alias{listExistingObjects}
\title{Get the names of the Objects in a table that is stored in a SQLite database}
\usage{
listExistingObjects(tableName, database)
}
\arguments{
\item{tableName}{table to append to}

\item{database}{The Path to the database that stores the information}
}
\value{
A character vector of objects stored in the table
}
\description{
Get the names of the Objects in a table that is stored in a SQLite database
}
\examples{
\dontrun{
exampleVector<-c(1,2,3,4)
appendObjectToTable(exampleVector, "Example_Vector", "ListOfVectors","Example.DB")

objectNameVector<-listExistingObjects("ListOfVectors","Example.DB")

identical(objectNameVector,"Example_Vector")
}

}
\seealso{
Other listAvailable: \code{\link{listExistingObjectsByTable}},
  \code{\link{listExistingTables}},
  \code{\link{nSerializedObjects}}
}
