% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAvailable.R
\name{listExistingObjectsByTable}
\alias{listExistingObjectsByTable}
\title{Get the names of the Objects by Axioserializer compliant tables in a SQLite database}
\usage{
listExistingObjectsByTable(database)
}
\arguments{
\item{database}{The Path to the database that stores the information}
}
\value{
A data.frame with two columns: Table and Objects
}
\description{
Get the names of the Objects by Axioserializer compliant tables in a SQLite database
}
\examples{
\dontrun{
exampleVector<-c(1,2,3,4)
appendObjectToTable(exampleVector, "Example_Vector", "ListOfVectors","Example.DB")

ObjectTableDF<-listExistingObjectsByTable("Example.DB")

identical(ObjectTableDF,data.frame(Table="ListOfVectors",ObjectName="Example_Vector")
}

}
\seealso{
Other listAvailable: \code{\link{listExistingObjects}},
  \code{\link{listExistingTables}},
  \code{\link{nSerializedObjects}}
}
