% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAvailable.R
\name{listExistingTables}
\alias{listExistingTables}
\title{Get the names of the Axioserializer compliant tables in a SQLite database}
\usage{
listExistingTables(database)
}
\arguments{
\item{database}{The Path to the database that stores the information}
}
\value{
A character vector of objects stored in the table
}
\description{
Get the names of the Axioserializer compliant tables in a SQLite database
}
\examples{
\dontrun{
exampleVector<-c(1,2,3,4)
appendObjectToTable(exampleVector, "Example_Vector", "ListOfVectors","Example.DB")

tableVector<-listExistingTables("Example.DB")

identical(tableVector,"ListOfVectors")
}

}
\seealso{
Other listAvailable: \code{\link{listExistingObjectsByTable}},
  \code{\link{listExistingObjects}},
  \code{\link{nSerializedObjects}}
}
