% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAvailable.R
\name{nSerializedObjects}
\alias{nSerializedObjects}
\title{Get the number of the Objects in a table that is stored in a SQLite database}
\usage{
nSerializedObjects(tableName, database)
}
\arguments{
\item{tableName}{table to append to}

\item{database}{The Path to the database that stores the information}
}
\value{
A numeric
}
\description{
Get the number of the Objects in a table that is stored in a SQLite database
}
\examples{
\dontrun{
exampleVector<-c(1,2,3,4)
appendObjectToTable(exampleVector, "Example_Vector", "ListOfVectors","Example.DB")

nObjects<-nSerializedObjects("ListOfVectors","Example.DB")

identical(nObjects,1)
}

}
\seealso{
Other listAvailable: \code{\link{listExistingObjectsByTable}},
  \code{\link{listExistingObjects}},
  \code{\link{listExistingTables}}
}
