% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readObjectFromDB.R
\name{readObjectFromDB}
\alias{readObjectFromDB}
\title{Read stored Object from inside a SQLite database}
\usage{
readObjectFromDB(objectToRead, databaseToReadFrom, compressionType = "bzip2")
}
\arguments{
\item{objectToRead}{Any R object to be stored inside the database}

\item{databaseToReadFrom}{The Path to the database that the object will read from, must end in '.DB'}
}
\value{
an R object
}
\description{
Un-Serializes objects from the database, returning them in their original state
}
\examples{
\dontrun{
exampleVector<-c(1,2,3,4)
writeObjectToDB(exampleVector,"Example.DB")

exampleVector2<-readObjectFromDB("exampleVector","Example.DB")

identical(exampleVector,exampleVector2)
}

}
