% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readObjectFromTable.R
\name{readObjectFromTable}
\alias{readObjectFromTable}
\title{Read add an Object to a list of objects inside a SQLite database}
\usage{
readObjectFromTable(objectToRead, tableName, databaseToReadFrom,
  compressionType = "bzip2")
}
\arguments{
\item{tableName}{table to append to}

\item{objectToSave}{Any R object to be stored inside the database}

\item{ObjectName}{The name the object will be saved as in table. If left blank, will be the name of the object}

\item{databaseToWriteTo}{The Path to the database that the object will read from, must end in '.DB'}
}
\description{
Read add an Object to a list of objects inside a SQLite database
}
\examples{
\dontrun{
exampleVector<-c(1,2,3,4)
appendObjectToTable(exampleVector, "Example_Vector", "ListOfVectors","Example.DB")

exampleVector2<-readObjectFromTable("Example_Vector","ListOfVectors","Example.DB")

identical(exampleVector,exampleVector2)
}

}
