% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnMultipleObjectsFromDB.R
\name{returnMultipleObjectsFromDB}
\alias{returnMultipleObjectsFromDB}
\title{Return stored Object(s) from inside a SQLite database to the Global Environment}
\usage{
returnMultipleObjectsFromDB(databaseToReadFrom, objectNames = NULL,
  environmentToReturnTo = .GlobalEnv, compressionType = "bzip2")
}
\arguments{
\item{databaseToReadFrom}{The Path to the database that the object(s) will be read from, must end in '.DB'}

\item{objectNames}{Any R object to be stored inside the database. If null, returns All the objects stored in the database to the specified environment}

\item{environmentToReturnTo}{The name of an existing environment to return the objects to, must  be a string. Defaults to the Global Environment}
}
\value{
None
}
\description{
Un-Serializes object(s) from the database, returning them in their original state, and putting them back into a specified environment
}
\examples{
\dontrun{
exampleVector<-testVector<-c(1,2,3,4)
writeObjectToDB(exampleVector,"Example.DB")

rm(exampleVector)

returnMultipleObjectsFromDB("Example.DB")

identical(exampleVector,testVector)
}

}
