% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeObjectToDB.R
\name{writeObjectToDB}
\alias{writeObjectToDB}
\title{Store any Object inside a SQLite database}
\usage{
writeObjectToDB(objectToSave, databaseToWriteTo, tableName = NULL,
  overwriteEntry = FALSE, compressionType = "bzip2")
}
\arguments{
\item{objectToSave}{Any R object to be stored inside the database}

\item{databaseToWriteTo}{The Path to the database that the object will be put into, must end in '.DB'}

\item{tableName}{The name the object will be saved in. If left blank, will be the name of the object}

\item{overwriteEntry}{Do you overwrite the entry in the database? Defaults to FALSE. Will throw an error if object already exists.}
}
\value{
None
}
\description{
Serializes any class of object and stores them in a database table, with the objects name as the table name
}
\examples{
\dontrun{
exampleVector<-c(1,2,3,4)
writeObjectToDB(exampleVector,"Example.DB")
}

}
