% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeObjectToTable.R
\name{writeObjectToTable}
\alias{writeObjectToTable}
\title{Read add an Object to a list of objects inside a SQLite database}
\usage{
writeObjectToTable(objectToSave, ObjectName, tableName, databaseToWriteTo,
  overwriteEntry = FALSE, compressionType = "bzip2",
  lockFile = tempfile())
}
\arguments{
\item{objectToSave}{Any R object to be stored inside the database}

\item{ObjectName}{The name the object will be saved as in table. If left blank, will be the name of the object}

\item{tableName}{table to append to}

\item{databaseToWriteTo}{The Path to the database that the object will read from, must end in '.DB'}
}
\description{
Read add an Object to a list of objects inside a SQLite database
}
\examples{
\dontrun{
exampleVector<-c(1,2,3,4)
appendObjectToTable(exampleVector, "Example_Vector", "ListOfVectors","Example.DB")

exampleVector2<-readObjectFromTable("Example_Vector","ListOfVectors","Example.DB")

identical(exampleVector,exampleVector2)
}

}
