% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readObjectFromTable.R
\name{readObjectFromTable}
\alias{readObjectFromTable}
\title{Read add an Object from inside a SQLite database table}
\usage{
readObjectFromTable(objectToRead, tableName, databaseToReadFrom, ...)
}
\arguments{
\item{objectToRead}{The name of theR object to be read from inside the database}

\item{tableName}{table to read from}

\item{databaseToReadFrom}{The Path to the database that the object will read from}

\item{...}{additional arguments to be passed to the decompression function \code{\link[base]{memDecompress}}, unlikely necessary}
}
\description{
Read add an Object from inside a SQLite database table
}
\examples{
\dontrun{
exampleVector<-c(1,2,3,4)
appendObjectToTable(exampleVector, "Example_Vector", "Example_Table","Example.DB")

exampleVector2<-readObjectFromTable("Example_Vector","Example_Table","Example.DB")

identical(exampleVector,exampleVector2)
}

}
